requirements.txt
PyPDF2==3.0.1
nltk==3.7
numpy==1.21.6
faiss-cpu==1.7.4
torch==1.13.1
transformers==4.26.1
sentence-transformers==2.2.2
langchain==0.0.150